+++ Synologen.Presentation.Intranet/Wpc/Synologen/ContractSales/SettlementList.ascx.cs
@@ -16,15 +16,17 @@ namespace Spinit.Wpc.Synologen.Presentation.Intranet.Wpc.Synologen.ContractSales
 			}
 		}
 
-		private void StorePageUrl() {
+		private void StorePageUrl() 
+		{
 			SynologenSessionContext.SettlementListPage = Request.Url.PathAndQuery;
 		}
 
-		private void PopulateSettlements() {
+		private void PopulateSettlements() 
+		{
 			if(MemberShopId<=0) return;
-			rptSettlements.DataSource = Provider.GetSettlementsDataSet(0,MemberShopId, null);
+			rptSettlements.DataSource = Provider.GetSettlementsDataSet(0, (int) MemberShopId, null);
 			rptSettlements.DataBind();
-			_shopNumber = Provider.GetShop(MemberShopId).Number;
+			_shopNumber = Provider.GetShop((int) MemberShopId).Number;
 			rptSettlements.Visible = (rptSettlements.Items.Count > 0);
 		}
